Reorganized code Removed the custom scheduler 'QueuedTaskScheduler' and replace it with the default one Added an option to delay lock re-acquisition in the DistributedLockManager
