+++ src/Reddoxx.Quartz.MongoDbJobStore/MongoDbJobStore.cs
@@ -293,19 +293,6 @@ public partial class MongoDbJobStore : IJobStore
         throw new NotSupportedException();
     }
 
-
-    public Task StoreJob(IJobDetail newJob, bool replaceExisting, CancellationToken token = default)
-    {
-        try
-        {
-            return ExecuteInTx(QuartzLockType.TriggerAccess, () => StoreJobInternal(newJob, replaceExisting), token);
-        }
-        catch (Exception ex)
-        {
-            throw new JobPersistenceException(ex.Message, ex);
-        }
-    }
-
     public Task StoreJobsAndTriggers(
         IReadOnlyDictionary<IJobDetail, IReadOnlyCollection<ITrigger>> triggersAndJobs,
         bool replace,
@@ -346,53 +333,6 @@ public partial class MongoDbJobStore : IJobStore
         {
             throw new JobPersistenceException(ex.Message, ex);
         }
-    }
-
-
-    public Task<bool> RemoveJob(JobKey jobKey, CancellationToken token = default)
-    {
-        try
-        {
-            return ExecuteInTx(QuartzLockType.TriggerAccess, () => RemoveJobInternal(jobKey), token);
-        }
-        catch (Exception ex)
-        {
-            throw new JobPersistenceException(ex.Message, ex);
-        }
-    }
-
-    public Task<bool> RemoveJobs(IReadOnlyCollection<JobKey> jobKeys, CancellationToken token = default)
-    {
-        try
-        {
-            return ExecuteInTx(
-                QuartzLockType.TriggerAccess,
-                async () =>
-                {
-                    var result = true;
-
-                    foreach (var jobKey in jobKeys)
-                    {
-                        result = result && await RemoveJobInternal(jobKey);
-                    }
-
-                    return result;
-                },
-                token
-            );
-        }
-        catch (Exception ex)
-        {
-            throw new JobPersistenceException(ex.Message, ex);
-        }
-    }
-
-
-    public async Task<IJobDetail?> RetrieveJob(JobKey jobKey, CancellationToken token = default)
-    {
-        var result = await _jobDetailRepository.GetJob(jobKey);
-
-        return result?.GetJobDetail();
     }
 
 
@@ -456,71 +396,11 @@ public partial class MongoDbJobStore : IJobStore
         return await _jobDetailRepository.GetJobGroupNames();
     }
 
-
     public async Task<IReadOnlyCollection<string>> GetTriggerGroupNames(CancellationToken token = default)
     {
         return await _triggerRepository.GetTriggerGroupNames();
     }
 
-
-    public Task PauseJob(JobKey jobKey, CancellationToken token = default)
-    {
-        try
-        {
-            return ExecuteInTx(
-                QuartzLockType.TriggerAccess,
-                async () =>
-                {
-                    var triggers = await GetTriggersForJob(jobKey, token);
-
-                    foreach (var operableTrigger in triggers)
-                    {
-                        await PauseTriggerInternal(operableTrigger.Key);
-                    }
-                },
-                token
-            );
-        }
-        catch (Exception ex)
-        {
-            throw new JobPersistenceException(ex.Message, ex);
-        }
-    }
-
-
-    public Task<IReadOnlyCollection<string>> PauseJobs(GroupMatcher<JobKey> matcher, CancellationToken token = default)
-    {
-        try
-        {
-            return ExecuteInTx<IReadOnlyCollection<string>>(
-                QuartzLockType.TriggerAccess,
-                async () =>
-                {
-                    var jobKeys = await _jobDetailRepository.GetJobsKeys(matcher);
-
-                    var groupNames = new HashSet<string>();
-                    foreach (var jobKey in jobKeys)
-                    {
-                        var triggers = await _triggerRepository.GetTriggers(jobKey);
-
-                        foreach (var trigger in triggers)
-                        {
-                            await PauseTriggerInternal(trigger.GetTriggerKey());
-                        }
-
-                        groupNames.Add(jobKey.Group);
-                    }
-
-                    return groupNames;
-                },
-                token
-            );
-        }
-        catch (Exception ex)
-        {
-            throw new JobPersistenceException(ex.Message, ex);
-        }
-    }
 
     public async Task<IReadOnlyCollection<string>> GetPausedTriggerGroups(CancellationToken cancellationToken = default)
     {
@@ -667,57 +547,6 @@ public partial class MongoDbJobStore : IJobStore
         }
     }
 
-    private async Task<bool> ReplaceTriggerInternal(TriggerKey triggerKey, IOperableTrigger newTrigger)
-    {
-        // SELECT
-        //  J.JOB_NAME,
-        //  J.JOB_GROUP,
-        //  J.IS_DURABLE,
-        //  J.JOB_CLASS_NAME,
-        //  J.REQUESTS_RECOVERY
-        // FROM
-        //  TRIGGERS T,
-        //  JOB_DETAILS J
-        // WHERE
-        //  T.SCHED_NAME = @schedulerName AND
-        //  T.SCHED_NAME = J.SCHED_NAME AND
-        //  T.TRIGGER_NAME = @triggerName AND
-        //  T.TRIGGER_GROUP = @triggerGroup AND
-        //  T.JOB_NAME = J.JOB_NAME AND
-        //  T.JOB_GROUP = J.JOB_GROUP";
-
-
-        var trigger = await _triggerRepository.GetTrigger(triggerKey);
-        if (trigger == null)
-        {
-            return false;
-        }
-
-        var result = await _jobDetailRepository.GetJob(trigger.JobKey);
-        var job = result?.GetJobDetail();
-
-        if (job == null)
-        {
-            return false;
-        }
-
-        if (!newTrigger.JobKey.Equals(job.Key))
-        {
-            throw new JobPersistenceException("New trigger is not related to the same job as the old trigger.");
-        }
-
-        var removedTrigger = await _triggerRepository.DeleteTrigger(triggerKey);
-        await StoreTriggerInternal(newTrigger, job, false, LocalTriggerState.Waiting, false, false);
-        return removedTrigger > 0;
-    }
-
-    private async Task<bool> RemoveJobInternal(JobKey jobKey)
-    {
-        await _triggerRepository.DeleteTriggers(jobKey);
-
-        var result = await _jobDetailRepository.DeleteJob(jobKey);
-        return result > 0;
-    }
 
     private async Task ResumeAllInternal()
     {
@@ -728,138 +557,6 @@ public partial class MongoDbJobStore : IJobStore
         );
 
         await _pausedTriggerGroupRepository.DeletePausedTriggerGroup(AllGroupsPaused);
-    }
-
-    private async Task StoreJobInternal(IJobDetail newJob, bool replaceExisting)
-    {
-        var existingJob = await _jobDetailRepository.JobExists(newJob.Key);
-
-        var jobDetail = new JobDetail(newJob, InstanceName);
-
-        if (existingJob)
-        {
-            if (!replaceExisting)
-            {
-                throw new ObjectAlreadyExistsException(newJob);
-            }
-
-            await _jobDetailRepository.UpdateJob(jobDetail);
-        }
-        else
-        {
-            await _jobDetailRepository.AddJob(jobDetail);
-        }
-    }
-
-    private async Task TriggeredJobCompleteInternal(
-        IOperableTrigger trigger,
-        IJobDetail jobDetail,
-        SchedulerInstruction triggerInstCode
-    )
-    {
-        try
-        {
-            switch (triggerInstCode)
-            {
-                case SchedulerInstruction.DeleteTrigger:
-                {
-                    if (!trigger.GetNextFireTimeUtc()
-                                .HasValue)
-                    {
-                        // double check for possible reschedule within job
-                        // execution, which would cancel the need to delete...
-                        var trig = await _triggerRepository.GetTrigger(trigger.Key);
-                        if (trig != null && !trig.NextFireTime.HasValue)
-                        {
-                            await RemoveTriggerInternal(trigger.Key, jobDetail);
-                        }
-                    }
-                    else
-                    {
-                        await RemoveTriggerInternal(trigger.Key, jobDetail);
-                        SignalSchedulingChangeOnTxCompletion(SchedulingSignalDateTime);
-                    }
-
-                    break;
-                }
-                case SchedulerInstruction.SetTriggerComplete:
-                {
-                    await _triggerRepository.UpdateTriggerState(trigger.Key, LocalTriggerState.Complete);
-                    SignalSchedulingChangeOnTxCompletion(SchedulingSignalDateTime);
-                    break;
-                }
-                case SchedulerInstruction.SetTriggerError:
-                {
-                    _logger.LogInformation("Trigger {Key} set to ERROR state.", trigger.Key);
-                    await _triggerRepository.UpdateTriggerState(trigger.Key, LocalTriggerState.Error);
-                    SignalSchedulingChangeOnTxCompletion(SchedulingSignalDateTime);
-                    break;
-                }
-                case SchedulerInstruction.SetAllJobTriggersComplete:
-                {
-                    await _triggerRepository.UpdateTriggersStates(trigger.JobKey, LocalTriggerState.Complete);
-                    SignalSchedulingChangeOnTxCompletion(SchedulingSignalDateTime);
-                    break;
-                }
-                case SchedulerInstruction.SetAllJobTriggersError:
-                {
-                    _logger.LogInformation("All triggers of Job {JobKey} set to ERROR state.", trigger.JobKey);
-                    await _triggerRepository.UpdateTriggersStates(trigger.JobKey, LocalTriggerState.Error);
-                    SignalSchedulingChangeOnTxCompletion(SchedulingSignalDateTime);
-                    break;
-                }
-            }
-
-            if (jobDetail.ConcurrentExecutionDisallowed)
-            {
-                await _triggerRepository.UpdateTriggersStates(
-                    jobDetail.Key,
-                    LocalTriggerState.Waiting,
-                    LocalTriggerState.Blocked
-                );
-
-                await _triggerRepository.UpdateTriggersStates(
-                    jobDetail.Key,
-                    LocalTriggerState.Paused,
-                    LocalTriggerState.PausedBlocked
-                );
-                SignalSchedulingChangeOnTxCompletion(SchedulingSignalDateTime);
-            }
-
-            if (jobDetail.PersistJobDataAfterExecution && jobDetail.JobDataMap.Dirty)
-            {
-                await _jobDetailRepository.UpdateJobData(jobDetail.Key, jobDetail.JobDataMap);
-            }
-        }
-        catch (Exception ex)
-        {
-            throw new JobPersistenceException(ex.Message, ex);
-        }
-
-        try
-        {
-            await _firedTriggerRepository.DeleteFiredTrigger(trigger.FireInstanceId);
-        }
-        catch (Exception ex)
-        {
-            throw new JobPersistenceException(ex.Message, ex);
-        }
-    }
-
-    protected virtual void SignalSchedulingChangeOnTxCompletion(DateTimeOffset candidateNewNextFireTime)
-    {
-        var sigTime = LogicalThreadContext.GetData<DateTimeOffset?>(KeySignalChangeForTxCompletion);
-        if (sigTime == null)
-        {
-            LogicalThreadContext.SetData(KeySignalChangeForTxCompletion, candidateNewNextFireTime);
-        }
-        else
-        {
-            if (candidateNewNextFireTime < sigTime)
-            {
-                LogicalThreadContext.SetData(KeySignalChangeForTxCompletion, candidateNewNextFireTime);
-            }
-        }
     }
 
     protected virtual DateTimeOffset? ClearAndGetSignalSchedulingChangeOnTxCompletion()
@@ -873,6 +570,7 @@ public partial class MongoDbJobStore : IJobStore
     {
         _schedulerSignaler.SignalSchedulingChange(candidateNewNextFireTime);
     }
+
 
     private void LogWarnIfNonZero(int val, [StructuredMessageTemplate] string? message, params object?[] args)
     {
